requirements.txt
Flask==1.0.3
requests==2.21.0
dialogflow==0.8.0
flask_cors==3.0.8
protobuf==3.11.3
python-dotenv==0.12.0
python_dateutil==2.8.1
requests_cache==0.5.2
